Split envelope model into smaller pieces
