+++ app/models/extensions/credential_registry_resources.rb
@@ -1,0 +1,12 @@
+require 'active_support/concern'
+
+# CredentialRegistry specific behavior for resource envelopes
+module CredentialRegistryResources
+  extend ActiveSupport::Concern
+
+  included do
+    def self.generate_ctid
+      "urn:guid:#{SecureRandom.uuid}"
+    end
+  end
+end
