+++ app/models/schema_config.rb
@@ -25,25 +25,6 @@ class SchemaConfig
     @prefix = prefix
   end
 
-  # List of all available configs
-  #
-  # Return: [List[String]] list of schema names
-  def self.all_configs
-    Dir['app/schemas/**/*.json']
-      .select { |path| !path.split('/').last.start_with?('_') }
-      .map    { |path| path.match(%r{app/schemas/(.*)/config.json})[1] }
-  end
-
-  # List of all available schemas
-  #
-  # Return: [List[String]] list of schema names
-  def self.all_schemas
-    Dir['app/schemas/**/*.json.erb']
-      .select { |path| !path.split('/').last.start_with?('_') }
-      .map    { |path| path.match(%r{app/schemas/(.*).json.erb})[1] }
-      .map    { |path| path.gsub(%r{/schema$}, '') }
-  end
-
   # resolve file paths for the corresponding schema name
   # Return: [String]
   def base_path
@@ -134,5 +115,49 @@ class SchemaConfig
   # schema validation. The public schema parse this to a public url
   def ref(name)
     "app/schemas/#{name}.json.erb"
+  end
+
+  # List of all available configs
+  #
+  # Return: [List[String]] list of schema names
+  def self.all_configs
+    Dir['app/schemas/**/*.json']
+      .select { |path| !path.split('/').last.start_with?('_') }
+      .map    { |path| path.match(%r{app/schemas/(.*)/config.json})[1] }
+  end
+
+  # List of all available schemas
+  #
+  # Return: [List[String]] list of schema names
+  def self.all_schemas
+    Dir['app/schemas/**/*.json.erb']
+      .select { |path| !path.split('/').last.start_with?('_') }
+      .map    { |path| path.match(%r{app/schemas/(.*).json.erb})[1] }
+      .map    { |path| path.gsub(%r{/schema$}, '') }
+  end
+
+  # get the resource_type for the envelope from the community config (if exists)
+  # Ex:
+  #   1) resource_type is a string
+  #   config: {"resource_type": "@type"}
+  #   processed_resource: "@type"='Bla'
+  #   >> 'Bla'
+  #
+  #   2) resource_type is an object with mapped values
+  #   config: {"resource_type": {
+  #             "property": "@type", "values_map": {"abc:Bla": 'bla'}
+  #            }}
+  #   processed_resource: "@type"='abc:Bla'
+  #   >> 'bla'
+  def self.resource_type_for(envelope)
+    cfg = new(envelope.community_name).config.try(:[], 'resource_type')
+    return nil if cfg.blank?
+
+    if cfg.is_a?(String)
+      envelope.processed_resource[cfg]
+    else
+      key = envelope.processed_resource[cfg['property']]
+      cfg['values_map'][key]
+    end
   end
 end
